Initial COWS libc implementation. Very experimental.

+++ basics/include/__typedef_ssize_t.h
@@ -2,6 +2,6 @@
 #define __wasm_sysroot___typedef_ssize_t_h
 
 /* This is defined to be the same size as size_t. */
-typedef long long ssize_t;
+typedef long ssize_t;
 
 #endif
